+++ crates/db/src/models/chatbans.rs
@@ -6,6 +6,6 @@ pub struct ChatBans {
     id: i32,
     user_id: i32,
     board_id: i32,
-    created_utc: i32,
+    created_utc: i64,
     banning_mod_id: i32
 }
