+++ crates/db/src/models/follows.rs
@@ -6,6 +6,6 @@ pub struct Follows {
     id: i32,
     user_id: i32,
     target_id: i32,
-    created_utc: i32,
+    created_utc: i64,
     get_notifs: bool,
 }
